Moved f2f and BAV functions into a new utility file
